Backend/requirements.txt
flask
psycopg2-binary
python-dotenv
flask-cors
pytest
requests
